refactor(mcp): split backend module and unify table/function specs
- Extract MCP client, fetch, function, provider, and tests into separate
  files under `coral-engine/src/backends/mcp/`.
- Introduce typed `McpTableSpec`/`McpTableFunctionSpec` wrappers around
  shared `TableCommon`/`SourceTableFunctionSpec`, with raw deserialize
  structs and explicit filter bindings.
