+++ crates/coral-spec/src/lib.rs
@@ -87,8 +87,8 @@ mod validate;
 
 pub use backends::http::{AuthSpec, BasicAuthSpec, CustomAuthSpec, HeaderAuthSpec};
 pub use backends::mcp::{
-    McpEnvSpec, McpLimitBinding, McpServerSpec, McpSourceManifest, McpTableFilterSpec,
-    McpTableFunctionSpec, McpTableSpec,
+    McpEnvSpec, McpLimitBinding, McpServerSpec, McpSourceManifest, McpTableFilterBinding,
+    McpTableFilterSpec, McpTableFunctionSpec, McpTableSpec,
 };
 pub(crate) use common::validate_test_queries;
 pub use common::{
